Update solidity-support.mdx
add gasleft() docs. It appears L1 gas is deducted before the contract has the ability to measure gas left. If someone has a link to the code where this occurs that would be helpful for me.

https://goerli.arbiscan.io/tx/0xaab3167237132bd51eaf14503a016ce764bd733145f2039035ffc45043af08ba#txninfo

This simple transaction logs `gasleft()` at the first possible moment as well as `tx.gasprice`. Because `gasleft()` is `6,784` at the time of execution, total gas used is `32,630` and L1/L2 gas respectively are `9,896` and `22,734` it appears the L1 gas is already deducted from the `gasleft()` precompile.

Gas limit - `gasleft()` 37,862 - 6,784 = 31,078 where 32,630 total gas is used means that the remainder of the function used 1,552 gas.

Contract used:

```
pragma solidity 0.8.16;

contract GasStation {
    event Stats(uint gasUsed, uint gasPrice);
    function printStats() external {
        emit Stats(gasleft(), tx.gasprice);
    }
}
```